+++ src/main/java/net/boreeas/riotapi/com/riotgames/platform/summoner/masterybook/MasteryEntry.java
@@ -37,5 +37,5 @@ public class MasteryEntry {
     @SerializedName(name = "talent")
     private Mastery mastery;
 
-    private double summonerId;
+    private long summonerId;
 }
