+++ src/main/java/net/boreeas/riotapi/com/riotgames/platform/summoner/runes/SummonerRune.java
@@ -33,5 +33,5 @@ public class SummonerRune {
     private int runeId;
     private int quantity;
     private Rune rune;
-    private double summonerId;
+    private long summonerId;
 }
